cart-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.8.0
python-dotenv==1.0.1
pydantic==2.9.2
python-jose[cryptography]==3.3.0

